worked on more endpoints in groups and events router, going to update…

+++ backend/db/migrations/20231001210837-create-Event.js
@@ -46,7 +46,7 @@ module.exports = {
         allowNull: false
       },
       price: {
-        type: Sequelize.INTEGER,
+        type: Sequelize.DECIMAL,
         allowNull: false
       },
       startDate: {
